Slight reorganisation and implement custom converters for slash commands
