+++ docs/source/api_references/parser.rst
@@ -1,0 +1,6 @@
+====================
+Parser API Reference
+====================
+
+.. automodule:: lightbulb.parser
+   :members:
