+++ lightbulb/utils/__init__.py
@@ -18,13 +18,11 @@
 from __future__ import annotations
 
 __all__ = [
-    "BaseParser",
     "ButtonNavigator",
     "ComponentButton",
     "DataStore",
     "EmbedPaginator",
     "Paginator",
-    "Parser",
     "ReactionButton",
     "ReactionNavigator",
     "StringPaginator",
@@ -38,6 +36,11 @@ __all__ = [
     "prev_page",
     "stop",
     "build_invite_url",
+    "data_store",
+    "nav",
+    "pag",
+    "permissions",
+    "search",
 ]
 
 import typing as t
@@ -48,13 +51,11 @@ import hikari
 from lightbulb.utils import data_store
 from lightbulb.utils import nav
 from lightbulb.utils import pag
-from lightbulb.utils import parser
 from lightbulb.utils import permissions
 from lightbulb.utils import search
 from lightbulb.utils.data_store import *
 from lightbulb.utils.nav import *
 from lightbulb.utils.pag import *
-from lightbulb.utils.parser import *
 from lightbulb.utils.permissions import *
 from lightbulb.utils.search import *
